requirements.txt
streamlit
numpy
pillow
pandas
plotly
opencv-python-headless
joblib
onnxruntime
